Implement create repository backend logic
When adding a repository it now checks if a restic repository exists at that path.
If yes it is ensured that the provided password is correct before adding it.
If no repository exists the GUI gives the option to create a new repository, which is then created with the given password.